refactor(ctxkeys): split Cell-model keys into kernel/ctxkeys
Move CellID, SliceID, JourneyID and their With/From helpers from
pkg/ctxkeys to a new kernel/ctxkeys package. pkg/ctxkeys now contains
only generic observability/networking keys (RequestID, TraceID, SpanID,
CorrelationID, RealIP). Each package keeps its own unexported ctxKey
type. Update runtime/observability/logging to dual-import both packages.

pkg/httputil and pkg/query continue to import pkg/ctxkeys for the
generic keys, preserving the pkg/ layer independence rule.

Refs: P0-2 PKG-CTXKEYS-LAYER-01
